src/SCAP/requirements.txt
altgraph==0.17.4
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
idna==3.10
packaging==24.1
paramiko==3.5.0
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyNaCl==1.5.0
pyspnego==0.11.1
pywinrm==0.5.0
requests==2.32.3
requests_ntlm==1.3.0
setuptools==75.1.0
urllib3==2.2.3
WMI==1.4.9
xmltodict==0.13.0
